Add UBSan workflow and fix/suppress misc. UBSan warnings. (#312)
* Fix UBSan enum errors in configure unit tests.
* Fix vfgetd UBSan warning.
* Fix misc. UB bugs in Robotic/expressions.
* Tweak --enable-ubsan flags, add UBSan workflow.
* Fix clang UBSan errors caused by math on NULL pointers in memfile.h.
* Temporary patch for libxmp list.h undefined behavior hack.

+++ src/configure.c
@@ -701,7 +701,7 @@ static void config_enable_oversampling(struct config_info *conf, char *name,
 {
   boolean result;
   if(config_boolean(&result, value))
-    conf->oversampling_on = (int)result;
+    conf->oversampling_on = result;
 }
 
 static void config_resample_mode(struct config_info *conf, char *name,
